Add rudimentary support for "arbitrary" dimensions in MultiThreshold
This allows node execution of MultiThreshold operators with arbitrary
number of dimensions, as long as the channel dimension is last. This is
necessary to run some verification steps of attention operators which,
at least for some intermediate steps, have 3 dimensional data layouts.

This does not change the behavior of execution on the already existing
2d and 4d data layouts.