[MultiThreshold] Generalize data layouts for node execution
The relevant aspect of the data layout annotation seems to be which axis
is labeled as the channel dimension "C": We do not actually have to care
about the total number and ordering of the other axes, as long as we can
find the index of the "C" axis and swap to have "C" at index 1 for node
execution (and swap it back afterwards).

Falls back to the default assumption that "C" is at index 1 if there is
no layout annotation, which is equivalent to the "NCHW" or "NC" layouts.

This is a rather experimental change which might break existing code and
is currently still restricted to the well-known 2-, 3- and 4-dimensional
layouts.